Fix 102-fibonacci.c file 1

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,7 +7,7 @@
  */
 int main(void)
 {
-	int i, j, k, next;
+	long int i, j, k, next;
 
 	j = 1;
 
@@ -17,10 +17,10 @@ int main(void)
 	{
 		if (j != 20365011074)
 		{
-			printf("%d, ", j);
+			printf("%ld, ", j);
 		} else
 		{
-			printf("%d\n", j);
+			printf("%ld\n", j);
 		}
 		next = j + k;
 		j = k;
